% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsamgeo-package.R
\docType{package}
\name{rsamgeo-package}
\alias{rsamgeo}
\alias{rsamgeo-package}
\title{rsamgeo: Segment Geospatial Data with the Segment Anything Model (SAM)}
\description{
Uses 'reticulate' to call 'Python' tools for segmenting geospatial data with the 'Segment Anything Model' (SAM). The 'segment-geospatial' package draws its inspiration from 'segment-anything-eo' repository authored by Aliaksandr Hancharenka.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brownag/rsamgeo}
  \item Report bugs at \url{https://github.com/brownag/rsamgeo/issues}
}

}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com}

}
\keyword{internal}
