% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samgeo.R
\name{sg_version}
\alias{sg_version}
\alias{samgeo}
\alias{sg_samgeo}
\title{\code{sg_version()}: Gets the \code{segment-geospatial} version}
\usage{
sg_version()

samgeo()

sg_samgeo(...)
}
\arguments{
\item{...}{Arguments to create model instance}
}
\value{
character. Version Number.

\code{sg_samgeo()}: return \code{SamGeo} model instance
}
\description{
Gets the \code{samgeo} module instance in use by the package in current \strong{R}/\code{reticulate} session.
}
