% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samgeo.R
\name{sg_generate}
\alias{sg_generate}
\title{Generate SamGeo Model Segmentation Output}
\usage{
sg_generate(x, ...)
}
\arguments{
\item{x}{A \code{SamGeo} model instance}

\item{...}{Arguments to \code{SamGeo.generate()}}
}
\description{
Create segmented image based on input GeoTIFF.
}
