% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samgeo.R
\name{sg_torch}
\alias{sg_torch}
\alias{sg_torch_cuda_is_available}
\title{Get \code{Module(torch)} Instance Used By \code{Module(samgeo)}}
\usage{
sg_torch()

sg_torch_cuda_is_available()
}
\value{
\code{Module(torch)} instance from \verb{Module(}

\code{sg_torch_cuda_is_available()}: \code{logical}. Is CUDA available?
}
\description{
Get \code{Module(torch)} Instance Used By \code{Module(samgeo)}
}
